requirements.txt
astroid==1.6.6
backports.functools-lru-cache==1.6.6
bcrypt==3.1.7
cffi==1.15.1
colorama==0.4.6
configparser==4.0.2
cryptography==3.3.2
dotenv==0.9.9
enum34==1.1.10
futures==3.4.0
ipaddress==1.0.23
isort==4.3.21
lazy-object-proxy==1.6.0
mccabe==0.6.1
paramiko==2.12.0
pycparser==2.21
pylint==1.9.5
PyNaCl==1.4.0
python-dotenv==0.18.0
robotframework==3.1.2
scp==0.13.6
setuptools==44.1.1
singledispatch==3.7.0
six==1.17.0
typing==3.10.0.0
# win-init-pton==1.1.0
wrapt==1.15.0
robotframework-sshlibrary==3.4.0
